+++ service/controller/clusternamespace/resource/clusternamespaceresources/delete.go
@@ -43,7 +43,7 @@ func (r *Resource) deleteRoleBinding(ctx context.Context, namespace string, role
 	} else if err != nil {
 		return microerror.Mask(err)
 	} else {
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("Deleting %#q roleBinding.", roleBinding))
+		r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("Deleting RoleBinding %#q in namespace %s.", roleBinding, namespace))
 
 		err = r.k8sClient.K8sClient().RbacV1().RoleBindings(namespace).Delete(ctx, roleBinding, metav1.DeleteOptions{})
 		if apierrors.IsNotFound(err) {
@@ -51,7 +51,7 @@ func (r *Resource) deleteRoleBinding(ctx context.Context, namespace string, role
 		} else if err != nil {
 			return microerror.Mask(err)
 		}
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("RoleBinding %#q has been deleted.", roleBinding))
+		r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("RoleBinding %#q in namespace %s has been deleted.", roleBinding, namespace))
 	}
 	return nil
 }
@@ -65,7 +65,7 @@ func (r *Resource) deleteRole(ctx context.Context, namespace string, role string
 	} else if err != nil {
 		return microerror.Mask(err)
 	} else {
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("Deleting %#q role.", role))
+		r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("Deleting Role %#q in namespace %s.", role, namespace))
 
 		err = r.k8sClient.K8sClient().RbacV1().Roles(namespace).Delete(ctx, role, metav1.DeleteOptions{})
 		if apierrors.IsNotFound(err) {
@@ -73,7 +73,7 @@ func (r *Resource) deleteRole(ctx context.Context, namespace string, role string
 		} else if err != nil {
 			return microerror.Mask(err)
 		}
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("Role %#q has been deleted.", role))
+		r.logger.LogCtx(ctx, "level", "info", "message", fmt.Sprintf("Role %#q in namespace %s has been deleted.", role, namespace))
 	}
 	return nil
 }